# packages/server/controllers/user.go
package controllers

import (
	"gardenai/server/database"
	"gardenai/server/models"
	"gardenai/server/validators"
	"strconv"

	"github.com/gofiber/fiber/v2"
	gonanoid "github.com/matoous/go-nanoid/v2"
	"golang.org/x/crypto/bcrypt"
)

type user struct{}

var User user

// Create godoc
// @Summary      Create a user
// @Description  Create a user with data
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        object   path    	object  true  "User JSON"
// @Success      200  {object}  user
// @Failure      400  {object}  error
// @Failure      404  {object}  error
// @Failure      500  {object}  error
// @Router       /user/signup/ [post]
func (user) Create(c *fiber.Ctx) error {
	user := new(validators.UserValidator)

	if err := c.BodyParser(user); err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"success": false,
			"reason":  err.Error(),
		})
	}

	err := validators.User.ValidateStruct(*user)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"success": false,
			"reason":  err.Error(),
		})
	}

	hashByte, err := bcrypt.GenerateFromPassword([]byte(user.Password), 14)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"success": false,
			"reason":  "Couldn't hash password"})
	}

	dbUser := models.User{
		Email:    user.Email,
		Password: string(hashByte),
	}

	id, err := gonanoid.Generate("0123456789", 9)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"success": false,
			"reason":  "Couldn't generate id"})
	} else {
		u64, err := strconv.ParseUint(id, 10, 32)
		if err != nil {
			return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
				"success": false,
				"reason":  "Couldn't parse id"})
		}

		dbUser.ID = uint(u64)
	}

	if err := database.DB.Create(&dbUser).Error; err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"success":   false,
			"reason":    "Couldn't create user",
			"db.reason": err.Error(),
		})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"success": true,
		"result":  dbUser.ID,
	})
}

// GetUserWithEmailAndPassword godoc
// @Summary      Get a user
// @Description  Get a user with user and password
// @Tags         user
// @Accept       json
// @Produce      json
// @Param        object   path    	object  true  "User JSON"
// @Success      200  {object}  user
// @Failure      400  {object}  error
// @Failure      404  {object}  error
// @Failure      500  {object}  error
// @Router       /user/signin/ [post]
func (user) GetUserWithEmailAndPassword(c *fiber.Ctx) error {
	user := new(validators.UserValidator)

	if err := c.BodyParser(user); err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"success": false,
			"reason":  err.Error(),
		})
	}

	err := validators.User.ValidateStruct(*user)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"success": false,
			"reason":  err.Error(),
		})
	}

	result := models.User{}
	if err := database.DB.Find(&result, "email = ?", user.Email).Error; err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"success":   false,
			"reason":    "Couldn't find user",
			"db.reason": err.Error(),
		})
	}

	if err := bcrypt.CompareHashAndPassword([]byte(result.Password), []byte(user.Password)); err != nil {
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
			"success": false,
			"reason":  "Invalid password",
		})
	}

	return c.JSON(fiber.Map{
		"success": true,
		"result": fiber.Map{
			"id":    result.ID,
			"email": result.Email,
		},
	})
}